requirements.txt.txt
streamlit
pytesseract
pillow
reportlab
groq
